backend/requirements.txt
fastapi==0.104.3
uvicorn==0.33.0
sqlalchemy==2.0.44
pydantic==2.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10